Move rust quote generation out to separate feature
