+++ untrusted/dcap/Cargo.toml
@@ -9,3 +9,9 @@ mc-sgx-urts = { path = "../urts" }
 
 [dev-dependencies]
 test_enclave = { path = "../test_enclave" }
+
+[features]
+default = []
+# TODO rebase this on https://github.com/mobilecoinfoundation/sgx/pull/14 to get
+#   the features from the `mc-sgx-dcap-sys` crate
+generate = []
