+++ untrusted/dcap/src/quote.rs
@@ -1,0 +1,4 @@
+/// A quote from the SGX interface
+pub struct Quote {
+    pub quote: Vec<u8>,
+}
